fix: 🐛 address code review findings for stable device identity
- move async_migrate_entry from ConfigFlow method to module-level
  function in new migration.py (HA calls module-level, not ConfigFlow)
- decouple device_slug from entry.unique_id — always use slugified
  user-chosen device name for suggested_object_id
- use CONF_HOST constant instead of hard-coded "host" in error dict
- fix docstring in async_get_device_serial to reflect actual 24-char
  hex format from 6 registers
- add cannot_read_modbus and already_configured translation strings
  to all 7 language files
- fix dead code in entity.py fallback chain (f"neopool_" is truthy)
- move logging/_LOGGER to module level in helpers.py
- restore TCP connection failure test (cannot_connect)
- add migration tests to test_init.py (success, serial unavailable,
  duplicate detection, entity update error)
- move migration tests from test_config_flow.py to test_init.py
