+++ custom_components/vistapool/config_flow.py
@@ -19,9 +19,8 @@ import logging
 
 import voluptuous as vol
 from homeassistant import config_entries
-from homeassistant.config_entries import ConfigEntry, ConfigFlowResult
+from homeassistant.config_entries import ConfigFlowResult
 from homeassistant.const import CONF_HOST, CONF_NAME, CONF_PORT
-from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers import translation as ha_translation
 from homeassistant.helpers.selector import SelectSelector, SelectSelectorConfig
 
@@ -74,89 +73,6 @@ class VistaPoolConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):  # type: ig
             return t.get(key) or DEFAULT_NAME
         except Exception:  # noqa: BLE001
             return DEFAULT_NAME
-
-    async def async_migrate_entry(self, config_entry: ConfigEntry) -> bool:
-        """Migrate config entry from v1 (no unique_id) to v2 (serial-based unique_id).
-
-        Old format: entry.unique_id = None
-        New format: entry.unique_id = "neopool_{serial}"
-        """
-        _LOGGER.info(
-            "Migrating VistaPool config entry %s from v1 to v2", config_entry.entry_id
-        )
-
-        # Trial read to get serial (dict() unwraps MappingProxy for the Modbus client)
-        serial = await async_get_device_serial(dict(config_entry.data))
-        if not serial:
-            _LOGGER.warning(
-                "Migration for %s: Cannot read device serial. "
-                "Will use entry_id as fallback identifier.",
-                config_entry.entry_id,
-            )
-            # Fallback: use entry_id as unique_id (not ideal, but safe)
-            new_unique_id = config_entry.entry_id
-        else:
-            new_unique_id = f"neopool_{serial}"
-
-        # Check if this serial is already registered (duplicate after migration)
-        if serial:
-            for entry in self.hass.config_entries.async_entries(DOMAIN):
-                if (
-                    entry.entry_id != config_entry.entry_id
-                    and entry.unique_id == new_unique_id
-                ):
-                    _LOGGER.error(
-                        "Migration failed: Device %s is already configured",
-                        new_unique_id,
-                    )
-                    return False
-
-        # Update config entry unique_id
-        self.hass.config_entries.async_update_entry(
-            config_entry, unique_id=new_unique_id
-        )
-
-        # Migrate entity unique_ids in registry
-        entity_registry = er.async_get(self.hass)
-        old_entry_id = config_entry.entry_id
-        migrated_count = 0
-
-        for entity_entry in list(
-            er.async_entries_for_config_entry(entity_registry, config_entry.entry_id)
-        ):
-            # Old format: {entry_id}_{key}
-            # New format: {new_unique_id}_{key}
-            if entity_entry.unique_id and entity_entry.unique_id.startswith(
-                old_entry_id
-            ):
-                key_part = entity_entry.unique_id.replace(f"{old_entry_id}_", "", 1)
-                migrated_unique_id = f"{new_unique_id}_{key_part}"
-
-                _LOGGER.debug(
-                    "Migrating entity %s: %s → %s",
-                    entity_entry.entity_id,
-                    entity_entry.unique_id,
-                    migrated_unique_id,
-                )
-
-                try:
-                    entity_registry.async_update_entity(
-                        entity_entry.entity_id,
-                        new_unique_id=migrated_unique_id,
-                    )
-                    migrated_count += 1
-                except Exception as err:
-                    _LOGGER.error(
-                        "Failed to migrate entity %s: %s", entity_entry.entity_id, err
-                    )
-
-        _LOGGER.info(
-            "Migration completed for %s: %d entities migrated, unique_id=%s",
-            config_entry.title,
-            migrated_count,
-            new_unique_id,
-        )
-        return True
 
     async def async_step_user(self, user_input=None) -> ConfigFlowResult:
         """Handle the initial step of the configuration flow."""
@@ -222,7 +138,7 @@ class VistaPoolConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):  # type: ig
             # Validation 2: Trial Modbus read → get serial number
             serial_number = await async_get_device_serial(user_input)
             if not serial_number:
-                errors["host"] = "cannot_read_modbus"
+                errors[CONF_HOST] = "cannot_read_modbus"
                 _LOGGER.warning(
                     "User cannot read from Modbus device at %s:%s",
                     user_input.get(CONF_HOST),
