+++ tests/test_config_flow.py
@@ -840,178 +840,6 @@ async def test_create_entry_with_duplicate_prevention():
 
 
 @pytest.mark.asyncio
-async def test_async_migrate_entry_v1_to_v2_success():
-    """Test migration from v1 (no unique_id) to v2 (serial-based unique_id)."""
-    flow = config_flow.VistaPoolConfigFlow()
-    flow.hass = MagicMock()
-
-    # Simulate v1 config entry without unique_id
-    config_entry = MagicMock()
-    config_entry.entry_id = "old_entry_id_123"
-    config_entry.unique_id = None
-    config_entry.title = "My Pool"
-    config_entry.data = {"host": "192.168.1.100", "port": DEFAULT_PORT, "slave_id": 1}
-
-    # No existing entries with this serial
-    flow.hass.config_entries.async_entries.return_value = []
-
-    # Mock entity registry with two old-format entities
-    mock_entity1 = MagicMock()
-    mock_entity1.unique_id = "old_entry_id_123_mbf_ph_measure"
-    mock_entity1.entity_id = "sensor.pool_ph"
-
-    mock_entity2 = MagicMock()
-    mock_entity2.unique_id = "old_entry_id_123_mbf_temperature"
-    mock_entity2.entity_id = "sensor.pool_temperature"
-
-    mock_registry = MagicMock()
-    mock_registry.async_update_entity = MagicMock()
-
-    expected_unique_id = f"neopool_{DEFAULT_SERIAL_STRING}"
-
-    with (
-        patch(
-            "custom_components.vistapool.config_flow.async_get_device_serial",
-            new=AsyncMock(return_value=DEFAULT_SERIAL_STRING),
-        ),
-        patch(
-            "custom_components.vistapool.config_flow.er.async_get",
-            return_value=mock_registry,
-        ),
-        patch(
-            "custom_components.vistapool.config_flow.er.async_entries_for_config_entry",
-            return_value=[mock_entity1, mock_entity2],
-        ),
-    ):
-        result = await flow.async_migrate_entry(config_entry)
-
-    assert result is True
-    # Verify unique_id was set on config entry
-    flow.hass.config_entries.async_update_entry.assert_called_once_with(
-        config_entry, unique_id=expected_unique_id
-    )
-    # Verify entity unique_ids were migrated
-    assert mock_registry.async_update_entity.call_count == 2
-    mock_registry.async_update_entity.assert_any_call(
-        "sensor.pool_ph",
-        new_unique_id=f"{expected_unique_id}_mbf_ph_measure",
-    )
-    mock_registry.async_update_entity.assert_any_call(
-        "sensor.pool_temperature",
-        new_unique_id=f"{expected_unique_id}_mbf_temperature",
-    )
-
-
-@pytest.mark.asyncio
-async def test_async_migrate_entry_v1_to_v2_serial_unavailable():
-    """Test migration falls back to entry_id when serial cannot be read."""
-    flow = config_flow.VistaPoolConfigFlow()
-    flow.hass = MagicMock()
-
-    config_entry = MagicMock()
-    config_entry.entry_id = "old_entry_id_456"
-    config_entry.unique_id = None
-    config_entry.title = "My Pool"
-    config_entry.data = {"host": "192.168.1.100", "port": DEFAULT_PORT, "slave_id": 1}
-
-    mock_registry = MagicMock()
-
-    with (
-        patch(
-            "custom_components.vistapool.config_flow.async_get_device_serial",
-            new=AsyncMock(return_value=None),
-        ),
-        patch(
-            "custom_components.vistapool.config_flow.er.async_get",
-            return_value=mock_registry,
-        ),
-        patch(
-            "custom_components.vistapool.config_flow.er.async_entries_for_config_entry",
-            return_value=[],
-        ),
-    ):
-        result = await flow.async_migrate_entry(config_entry)
-
-    assert result is True
-    # Falls back to entry_id when serial is unavailable
-    flow.hass.config_entries.async_update_entry.assert_called_once_with(
-        config_entry, unique_id="old_entry_id_456"
-    )
-
-
-@pytest.mark.asyncio
-async def test_async_migrate_entry_v1_to_v2_duplicate_detected():
-    """Test migration fails when another entry already has the same serial."""
-    flow = config_flow.VistaPoolConfigFlow()
-    flow.hass = MagicMock()
-
-    config_entry = MagicMock()
-    config_entry.entry_id = "entry_aaa"
-    config_entry.unique_id = None
-    config_entry.title = "My Pool"
-    config_entry.data = {"host": "192.168.1.100", "port": DEFAULT_PORT, "slave_id": 1}
-
-    # Another entry already has this serial
-    existing_entry = MagicMock()
-    existing_entry.entry_id = "entry_bbb"
-    existing_entry.unique_id = f"neopool_{DEFAULT_SERIAL_STRING}"
-    flow.hass.config_entries.async_entries.return_value = [existing_entry]
-
-    with patch(
-        "custom_components.vistapool.config_flow.async_get_device_serial",
-        new=AsyncMock(return_value=DEFAULT_SERIAL_STRING),
-    ):
-        result = await flow.async_migrate_entry(config_entry)
-
-    assert result is False
-    # async_update_entry should NOT have been called
-    flow.hass.config_entries.async_update_entry.assert_not_called()
-
-
-@pytest.mark.asyncio
-async def test_async_migrate_entry_entity_update_error():
-    """Test migration continues when individual entity update fails."""
-    flow = config_flow.VistaPoolConfigFlow()
-    flow.hass = MagicMock()
-
-    config_entry = MagicMock()
-    config_entry.entry_id = "old_entry_id_789"
-    config_entry.unique_id = None
-    config_entry.title = "My Pool"
-    config_entry.data = {"host": "192.168.1.100", "port": DEFAULT_PORT, "slave_id": 1}
-
-    flow.hass.config_entries.async_entries.return_value = []
-
-    # Entity whose update will raise an exception
-    mock_entity = MagicMock()
-    mock_entity.unique_id = "old_entry_id_789_mbf_ph_measure"
-    mock_entity.entity_id = "sensor.pool_ph"
-
-    mock_registry = MagicMock()
-    mock_registry.async_update_entity.side_effect = ValueError("registry conflict")
-
-    with (
-        patch(
-            "custom_components.vistapool.config_flow.async_get_device_serial",
-            new=AsyncMock(return_value=DEFAULT_SERIAL_STRING),
-        ),
-        patch(
-            "custom_components.vistapool.config_flow.er.async_get",
-            return_value=mock_registry,
-        ),
-        patch(
-            "custom_components.vistapool.config_flow.er.async_entries_for_config_entry",
-            return_value=[mock_entity],
-        ),
-    ):
-        result = await flow.async_migrate_entry(config_entry)
-
-    # Migration should still succeed despite entity update error
-    assert result is True
-    mock_registry.async_update_entity.assert_called_once()
-
-
-@pytest.mark.asyncio
 async def test_trial_modbus_read_general_exception():
     """Test that trial Modbus read handles non-timeout exceptions gracefully."""
     from custom_components.vistapool.helpers import async_get_device_serial
