+++ tests/test_init.py
@@ -20,9 +20,11 @@ from homeassistant.exceptions import ServiceValidationError
 from custom_components.vistapool import (
     _cleanup_removed_entities,
     _register_services,
+    async_migrate_entry,
     async_setup_entry,
     async_unload_entry,
 )
+from custom_components.vistapool.const import DEFAULT_PORT
 
 
 @pytest.mark.asyncio
@@ -747,3 +749,171 @@ def test_cleanup_no_orphans():
             _cleanup_removed_entities(hass, entry)
 
     mock_registry.async_remove.assert_not_called()
+
+
+# --- Migration tests ---
+
+DEFAULT_SERIAL_REGS = [0x0000, 0x0001, 0x00AC, 0x00CD, 0x0012, 0x0034]
+DEFAULT_SERIAL_STRING = "0000000100AC00CD00120034"
+
+
+@pytest.mark.asyncio
+async def test_async_migrate_entry_v1_to_v2_success():
+    """Test migration from v1 (no unique_id) to v2 (serial-based unique_id)."""
+    hass = MagicMock()
+
+    config_entry = MagicMock()
+    config_entry.entry_id = "old_entry_id_123"
+    config_entry.unique_id = None
+    config_entry.version = 1
+    config_entry.title = "My Pool"
+    config_entry.data = {"host": "192.168.1.100", "port": DEFAULT_PORT, "slave_id": 1}
+
+    hass.config_entries.async_entries.return_value = []
+
+    mock_entity1 = MagicMock()
+    mock_entity1.unique_id = "old_entry_id_123_mbf_ph_measure"
+    mock_entity1.entity_id = "sensor.pool_ph"
+
+    mock_entity2 = MagicMock()
+    mock_entity2.unique_id = "old_entry_id_123_mbf_temperature"
+    mock_entity2.entity_id = "sensor.pool_temperature"
+
+    mock_registry = MagicMock()
+    mock_registry.async_update_entity = MagicMock()
+
+    expected_unique_id = f"neopool_{DEFAULT_SERIAL_STRING}"
+
+    with (
+        patch(
+            "custom_components.vistapool.migration.async_get_device_serial",
+            new=AsyncMock(return_value=DEFAULT_SERIAL_STRING),
+        ),
+        patch(
+            "custom_components.vistapool.migration.er.async_get",
+            return_value=mock_registry,
+        ),
+        patch(
+            "custom_components.vistapool.migration.er.async_entries_for_config_entry",
+            return_value=[mock_entity1, mock_entity2],
+        ),
+    ):
+        result = await async_migrate_entry(hass, config_entry)
+
+    assert result is True
+    hass.config_entries.async_update_entry.assert_called_once_with(
+        config_entry, unique_id=expected_unique_id, version=2
+    )
+    assert mock_registry.async_update_entity.call_count == 2
+    mock_registry.async_update_entity.assert_any_call(
+        "sensor.pool_ph",
+        new_unique_id=f"{expected_unique_id}_mbf_ph_measure",
+    )
+    mock_registry.async_update_entity.assert_any_call(
+        "sensor.pool_temperature",
+        new_unique_id=f"{expected_unique_id}_mbf_temperature",
+    )
+
+
+@pytest.mark.asyncio
+async def test_async_migrate_entry_v1_to_v2_serial_unavailable():
+    """Test migration falls back to entry_id when serial cannot be read."""
+    hass = MagicMock()
+
+    config_entry = MagicMock()
+    config_entry.entry_id = "old_entry_id_456"
+    config_entry.unique_id = None
+    config_entry.version = 1
+    config_entry.title = "My Pool"
+    config_entry.data = {"host": "192.168.1.100", "port": DEFAULT_PORT, "slave_id": 1}
+
+    mock_registry = MagicMock()
+
+    with (
+        patch(
+            "custom_components.vistapool.migration.async_get_device_serial",
+            new=AsyncMock(return_value=None),
+        ),
+        patch(
+            "custom_components.vistapool.migration.er.async_get",
+            return_value=mock_registry,
+        ),
+        patch(
+            "custom_components.vistapool.migration.er.async_entries_for_config_entry",
+            return_value=[],
+        ),
+    ):
+        result = await async_migrate_entry(hass, config_entry)
+
+    assert result is True
+    hass.config_entries.async_update_entry.assert_called_once_with(
+        config_entry, unique_id="old_entry_id_456", version=2
+    )
+
+
+@pytest.mark.asyncio
+async def test_async_migrate_entry_v1_to_v2_duplicate_detected():
+    """Test migration fails when another entry already has the same serial."""
+    hass = MagicMock()
+
+    config_entry = MagicMock()
+    config_entry.entry_id = "entry_aaa"
+    config_entry.unique_id = None
+    config_entry.version = 1
+    config_entry.title = "My Pool"
+    config_entry.data = {"host": "192.168.1.100", "port": DEFAULT_PORT, "slave_id": 1}
+
+    existing_entry = MagicMock()
+    existing_entry.entry_id = "entry_bbb"
+    existing_entry.unique_id = f"neopool_{DEFAULT_SERIAL_STRING}"
+    hass.config_entries.async_entries.return_value = [existing_entry]
+
+    with patch(
+        "custom_components.vistapool.migration.async_get_device_serial",
+        new=AsyncMock(return_value=DEFAULT_SERIAL_STRING),
+    ):
+        result = await async_migrate_entry(hass, config_entry)
+
+    assert result is False
+    hass.config_entries.async_update_entry.assert_not_called()
+
+
+@pytest.mark.asyncio
+async def test_async_migrate_entry_entity_update_error():
+    """Test migration continues when individual entity update fails."""
+    hass = MagicMock()
+
+    config_entry = MagicMock()
+    config_entry.entry_id = "old_entry_id_789"
+    config_entry.unique_id = None
+    config_entry.version = 1
+    config_entry.title = "My Pool"
+    config_entry.data = {"host": "192.168.1.100", "port": DEFAULT_PORT, "slave_id": 1}
+
+    hass.config_entries.async_entries.return_value = []
+
+    mock_entity = MagicMock()
+    mock_entity.unique_id = "old_entry_id_789_mbf_ph_measure"
+    mock_entity.entity_id = "sensor.pool_ph"
+
+    mock_registry = MagicMock()
+    mock_registry.async_update_entity.side_effect = ValueError("registry conflict")
+
+    with (
+        patch(
+            "custom_components.vistapool.migration.async_get_device_serial",
+            new=AsyncMock(return_value=DEFAULT_SERIAL_STRING),
+        ),
+        patch(
+            "custom_components.vistapool.migration.er.async_get",
+            return_value=mock_registry,
+        ),
+        patch(
+            "custom_components.vistapool.migration.er.async_entries_for_config_entry",
+            return_value=[mock_entity],
+        ),
+    ):
+        result = await async_migrate_entry(hass, config_entry)
+
+    assert result is True
+    mock_registry.async_update_entity.assert_called_once()
